fix(OutputView, LottoService): 로또 수익률 타입 변경
int, float 사용 시 수익률을 계산하는 과정에서 오버플로우 발생 -> long, double로 타입 변환

+++ src/main/java/lotto/service/LottoServiceImpl.java
@@ -65,13 +65,13 @@ public class LottoServiceImpl implements LottoService {
     }
 
     @Override
-    public float calculateRateOfReturn(Map<LottoRank, Integer> lottoResult, int price) {
-        int totalPrize = 0;
+    public double calculateRateOfReturn(Map<LottoRank, Integer> lottoResult, int price) {
+        long totalPrize = 0;
         for(LottoRank rank : LottoRank.values()) {
-            totalPrize += lottoResult.getOrDefault(rank, 0) * rank.getPrize();
+            totalPrize += (long) lottoResult.getOrDefault(rank, 0) * rank.getPrize();
         }
 
-        float rate = ((float) totalPrize / price) * 100f;
+        double rate = ((float) totalPrize / price) * 100f;
         return Math.round(rate * 100f) / 100f;
     }
